Create KML_File_Output.py
This code takes IP as input and creates a KML file which can be then uploaded to Google Earth. 
Scenario: While investigating an infected machine we came up with IP's it interacted with. This code helps in putting a placemark on google earth for all the IP's associated, helping in the better understanding of connections.