Create VulnerabilityScanner101.py
A simple vulnerability scanner python code. It connects to IP address/Network mentioned and figures out on the basis of services running whether the Target is running a vulnerable service or not.  Takes input of vulnerable banners from a .txt file. 
Specify the IP range, Ports we want to scan and a text file containing latest vulnerable services.